+++ app/commands/homebrew_context/daggerheart/copy_item_command.rb
@@ -21,7 +21,7 @@ module HomebrewContext
           input[:item].attributes
             .slice('kind', 'info')
             .symbolize_keys
-            .merge({ user: input[:user], name: input[:item].name['en'] })
+            .merge({ user: input[:user], name: input[:item].name['en'], description: input[:item].description['en'] })
         )[:result]
 
         { result: result }